dashboard-website/mortgage_dashboard/requirements.txt
Babel==2.9.1
django-bootstrap-v5==1.0.11
djangorestframework==3.13.1
django-cors-headers

